separated addon updater
